pyproject.toml
[tool.poetry]
name = "news-mcp-server"
version = "0.1.0"
description = "MCP server that searches news, fetches articles, and enables AI discussions via OpenAI, Gemini, or Ollama."
authors = ["Gizmo <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "news_mcp_server", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
mcp = "^1.2.1"
httpx = "^0.27.2"
feedparser = "^6.0.11"
trafilatura = "^1.8.0"
openai = "^1.54.3"
google-generativeai = "^0.8.2"
ollama = "^0.3.3"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
typing-extensions = "^4.12.2"

[tool.poetry.scripts]
news-mcp-server = "news_mcp_server.server:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
